% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_calcula_ajuste_2022.R
\encoding{UTF-8}
\name{epe4md_calcula_ajuste_2022}
\alias{epe4md_calcula_ajuste_2022}
\title{Realiza projeção assim como a função \link{epe4md_calcula}, porém com
ajuste em 2022 em função do impacto da Lei n° 14.300}
\usage{
epe4md_calcula_ajuste_2022(
  premissas_reg,
  ano_base,
  ano_max_resultado = 2050,
  altera_sistemas_existentes = FALSE,
  ano_decisao_alteracao = 2023,
  inflacao = 0.0375,
  taxa_desconto_nominal = 0.13,
  custo_reforco_rede = 200,
  ano_troca_inversor = 11,
  pagamento_disponibilidade = 0.3,
  disponibilidade_kwh_mes = 100,
  filtro_renda_domicilio = "maior_3sm",
  desconto_capex_local = 0,
  anos_desconto = 0,
  tx_cresc_grupo_a = 0.016,
  spb = 0.3,
  p_max = 0.01,
  q_max = 1,
  filtro_comercial = NA,
  ajuste_ano_corrente = FALSE,
  base_ano_corrente = NA,
  ultimo_mes_ajuste = NA,
  dir_dados_premissas = "inst/dados_premissas"
)
}
\arguments{
\item{premissas_reg}{data.frame. Input de premissas regulatórias para serem
consideradas nos cálculos.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{ano_max_resultado}{numeric. Ano final para apresentação dos resultados.
Máximo igual a 2050. Default igual a 2050.}

\item{altera_sistemas_existentes}{logic. TRUE se alterações regulatórias
afetam investimentos realizados em anos anteriores à revisão da regulação.
Default igual a FALSE.}

\item{ano_decisao_alteracao}{numeric. Ano em que são definidas novas regras
e se tornam de conhecimento público. Esse parâmetro só tem efeito caso o
anterior seja igual a TRUE. Default igual a 2023.}

\item{inflacao}{mumeric. Taxa anual de inflacao considerada no reajuste das
tarifas e para calcular o retorno real de projetos. Default igual a 0.0375.}

\item{taxa_desconto_nominal}{numeric. Taxa de desconto nominal considerada
nos cálculos de payback descontado. Default igual a 0.13.}

\item{custo_reforco_rede}{numeric. Custo em R$/kW aplicado a projetos de
geracao remota em Alta Tensão. Representa um custo pago pelo empreendedor
para reforços na rede. Default igual a 200.}

\item{ano_troca_inversor}{numeric. Ano, a partir do ano de instalação, em
que é realizada a troca do inversor fotovoltaico. Default igual a 11.}

\item{disponibilidade_kwh_mes}{numeric. Consumo de disponbilidade do
consumidor em kWh/mês. Default igual a 100, equivalente a um consumidor
trifásico.}

\item{filtro_renda_domicilio}{string. Define o filtro aplicado a consumidores
residenciais, de acordo com a renda mensal do responsável, em salários
mínimos. Permite: "maior_2sm", "maior_3sm" ou "maior_5sm". Default igual a
"maior_3sm".}

\item{desconto_capex_local}{numeric. Percentual de desconto a ser aplicado no
CAPEX de sistemas de geração local(ex: 0.1) para simulação de incentivos.
Default igual a 0.}

\item{anos_desconto}{vector. Anos em que há a incidência do desconto no
CAPEX. Default igual a 0.}

\item{ajuste_ano_corrente}{logic. Se TRUE indica que a projeção deverá
incorporar o histórico mensal recente, verificado em parte do primeiro ano
após o ano base. Default igual a FALSE.}

\item{base_ano_corrente}{data.frame Resultado da função
\link{epe4md_prepara_base} com histórico dos meses recentes.}

\item{ultimo_mes_ajuste}{numeric. Último mês com dados completos na
base_ano_corrente. @param dir_dados_premissas Diretório onde se encontram as
premissas. Se esse parâmetro não for passado, a função usa os dados default
que são instalados com o pacote. É importante que os nomes dos arquivos sejam
os mesmos da pasta default.}

\item{pagamento_disponibilidade.}{numeric. Percentual de meses em que o
consumidor residencial paga custo de disponbilidade em função da
variabilidade da geração FV. Default igual a 0.3.}
}
\value{
data.frame com os resultados da projeção de capacidade instalada
de micro e minigeração distribuída, número de adotantes e geração
mensal de energia.
}
\description{
Realiza projeção assim como a função \link{epe4md_calcula}, porém com
ajuste em 2022 em função do impacto da Lei n° 14.300
}
