% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_compara_backtesting.R
\encoding{UTF-8}
\name{epe4md_compara_backtesting}
\alias{epe4md_compara_backtesting}
\title{Calcula o erro da projeção em relação aos dados verificados.}
\usage{
epe4md_compara_backtesting(
  resultados_mensais,
  ano_base,
  ano_backtesting,
  dir_dados_premissas = NA_character_
)
}
\arguments{
\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas.
Se esse parâmetro não for passado, a função usa os dados default que são
instalados com o pacote. É importante que os nomes dos arquivos sejam os
mesmos da pasta default.}

\item{proj_mensal}{dataframe. Resultado da função
\link{epe4md_proj_mensal}.}
}
\value{
data.frame com os resultados da projeção de capacidade instalada
de micro e minigeração distribuída, número de adotantes e geração
mensal de energia.
}
\description{
Calcula o erro da projeção em relação aos dados verificados.
}
