% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_fatores_publicacao.R
\encoding{UTF-8}
\name{epe4md_fatores_publicacao}
\alias{epe4md_fatores_publicacao}
\title{Altera fatores para facilitar entendimento em publicações}
\usage{
epe4md_fatores_publicacao(dados)
}
\arguments{
\item{dados}{data.frame. Base de dados para ser alterada.}
}
\value{
data.frame com nomes dos segmentos alterados.
}
\description{
Altera fatores para facilitar entendimento em publicações
}
\examples{

dados <- tibble::tibble(
  fonte_resumo = c("comercial_bt",
                   "comercial_at_remoto",
                   "comercial_at",
                   "residencial_remoto",
                   "residencial")
)

fatores <- epe4md_fatores_publicacao(dados)
}
