% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_cap_arm_acum}
\alias{epe4md_graf_cap_arm_acum}
\title{Gráfico da capacidade de armazenamento acumulada}
\usage{
epe4md_graf_cap_arm_acum(
  dados,
  ano_inicio = 2013,
  cor = "#13475d",
  tamanho = 14
)
}
\arguments{
\item{dados}{data.frame Resultados mensais de potencia e energia}
}
\description{
Gráfico da capacidade de armazenamento acumulada
}
