% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_pot_bat_acum}
\alias{epe4md_graf_pot_bat_acum}
\title{Gráfico da capacidade instalada acumulada com baterias}
\usage{
epe4md_graf_pot_bat_acum(dados, ano_inicio = 2013, tamanho = 14)
}
\arguments{
\item{dados}{data.frame Resultados mensais de potencia e energia}
}
\description{
Gráfico da capacidade instalada acumulada com baterias
}
