% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_pot_cenario}
\alias{epe4md_graf_pot_cenario}
\title{Gr치fico da capacidade instalada acumulada por cen치rio}
\usage{
epe4md_graf_pot_cenario(dados, ano_inicio = 2013, tamanho = 14)
}
\arguments{
\item{dados}{data.frame Resultado resumido com coluna cenario}
}
\description{
Gr치fico da capacidade instalada acumulada por cen치rio
}
