% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_prepara_base.R
\encoding{UTF-8}
\name{epe4md_prepara_base}
\alias{epe4md_prepara_base}
\title{Prepara a base de dados da ANEEL com geradores existentes para ser utilizada
nas funções seguintes.}
\usage{
epe4md_prepara_base(
  base_aneel,
  ano_base,
  resumida = TRUE,
  dir_dados_premissas = NA_character_
)
}
\arguments{
\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a
função irá buscar a base de dados. Último ano completo realizado.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas. Se esse
parâmetro não for passado, a função usa os dados default que são instalados
com o pacote. É importante que os nomes dos arquivos sejam os mesmos da
pasta default.}

\item{base_aneel}{Dataframe com a base de dados disponibilizada pela ANEEL
na sua página de dados abertos com dados individualizados dos micro e
minigeradores distribuídos. Sugestão ler base com read.csv2}

\item{resumida}{Logic. Se TRUE, retorna a base resumida (Default). Se
FALSE retorna base com mais desagregações.}
}
\value{
data.frame. Base tratada e resumida.
}
\description{
Prepara a base de dados da ANEEL com geradores existentes para ser utilizada
nas funções seguintes.
}
\examples{
epe4md_prepara_base(
  base_aneel,
  ano_base = 2021,
  resumida = TRUE,
  dir_dados_premissas = NA_character_
)
}
