% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_proj_geracao.R
\encoding{UTF-8}
\name{epe4md_proj_geracao}
\alias{epe4md_proj_geracao}
\title{Estima a geração de eletricidade a partir da projeção de potência}
\usage{
epe4md_proj_geracao(
  proj_mensal,
  ano_base,
  filter_uf,
  dir_dados_premissas = "inst/\\n                                dados_premissas"
)
}
\arguments{
\item{proj_mensal}{dataframe. Resultado da função
\link{epe4md_proj_mensal}.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{filter_uf}{string. Parâmetro que define uma unidade federativa (UF) a ser
filtrada. Para um resultado com todas as unidades federativas, defina o parâmetro
como "N".}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas.
Se esse parâmetro não for passado, a função usa os dados default que são
instalados com o pacote. É importante que os nomes dos arquivos sejam os
mesmos da pasta default.}
}
\value{
data.frame com os resultados da projeção de capacidade instalada
de micro e minigeração distribuída, número de adotantes e geração
mensal de energia.
}
\description{
Estima a geração de eletricidade a partir da projeção de potência
}
\examples{

projecao_mensal <- tibble::tibble(
    ano = 2021,
    nome_4md = c("OUTRA", "RGE", "RORAIMA", "SULGIPE", "UHENPAL"),
    fonte_resumo = "Fotovoltaica",
    segmento = "comercial_at",
    adotantes_mes = c(4, 19, 2, 8, 5),
    pot_mes_mw = c(0.25500, 2.14532, 0.03060, 0.12604, 0.05441),
    mes_ano = "2021 jan",
    mes = 1,
    p = c(0.0004484362, 0.0004484362, 0.0001000000, 0.0001767164, 0.0036879405),
    q = c( 1.0000000, 1.0000000, 0.9044994, 0.8862171, 0.5501830),
    Ft = c(0.7847763, 0.8193315, 0.2750850, 0.3674506, 0.4915285)
)

epe4md_proj_geracao(
   proj_mensal = projecao_mensal,
   ano_base = 2021,
   dir_dados_premissas = NA_character_
)
}
