% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_proj_mensal.R
\encoding{UTF-8}
\name{epe4md_proj_mensal}
\alias{epe4md_proj_mensal}
\title{Faz a abertura mensal da projeção de potência}
\usage{
epe4md_proj_mensal(
  lista_potencia,
  ano_base,
  ano_max_resultado = 2050,
  ajuste_ano_corrente = FALSE,
  ultimo_mes_ajuste = NA,
  metodo_ajuste = NA,
  dir_dados_premissas = "inst/\\n                               dados_premissas"
)
}
\arguments{
\item{lista_potencia}{list. Resultado da função
\link{epe4md_proj_potencia}.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{ano_max_resultado}{numeric. Ano final para apresentação dos resultados.
Máximo igual a 2050. Default igual a 2050.}

\item{ajuste_ano_corrente}{logic. Se TRUE indica que a projeção deverá
incorporar o histórico mensal recente, verificado em parte do primeiro ano
após o ano base. Default igual a FALSE. O arquivo base_mmgd.xlsx deve
incorporar esse histórico.}

\item{ultimo_mes_ajuste}{numeric. Último mês com dados completos na
base_ano_corrente. Default igual a NA. Só tem efeito caso ajuste_ano_corrente
seja igual a TRUE.}

\item{metodo_ajuste}{string. Se igual a "extrapola" o modelo irá extrapolar a
potência e o número de adotantes até o final do ano base + 1 com base no
verificado até o ultimo_mes_ajuste. Se igual a "substitui", o modelo
substitui a projeção até o ultimo_mes_ajuste e mantém o restante do ano com a
projeção normal. Só tem efeito caso ajuste_ano_corrente seja igual a TRUE.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas.
Se esse parâmetro não for passado, a função usa os dados default que são
instalados com o pacote. É importante que os nomes dos arquivos sejam os
mesmos da pasta default.}
}
\value{
data.frame com os resultados da projeção de capacidade instalada
de micro e minigeração distribuída e número de adotantes mensal
}
\description{
Faz a abertura mensal da projeção de potência
}
\examples{

lista_potencia <- list(
    proj_potencia <- tibble(
      nome_4md = c("OUTRA", "RGE", "RORAIMA", "SULGIPE", "UHENPAL"),
      segmento = "comercial_at",
      p = c(0.0004484362, 0.0005572423, 0.000122345, 0.0002427430, 0.0001034085),
      q = 1,
      spb = 0.3,
      ano = 2021,
      Ft = c(0.7847763, 0.8193315, 0.4980477, 0.6633618, 0.4560837),
      consumidores = c(2801, 10995, 791, 116, 48),
      payback = c(9.300698, 7.295061, 11.896997, 12.448897, 11.590019),
      mercado_potencial = c(43.00, 308.00, 5.50, 0.75, 0.25),
      adotantes_ano = 0,
      adotantes_acum = c(0, 1, 0, 0, 0),
      fonte_resumo = "Eólica",
      part_fonte = c(0.0000000000, 0.0006309148, 0.0000000000, 0.0000000000, 0.0000000000),
      adotantes_hist = 0,
      pot_media = c(38.61667, 2.50000, 38.61667, 38.61667, 38.61667),
      pot_ano = 0,
      pot_hist = 0,
      pot_ano_mw = 0,
      pot_acum_mw = c(0.0000, 0.0025, 0.0000, 0.0000, 0.0000)
    ),
    part_adotantes <- tibble(
      ano = 2021,
      segmento = c("comercial_at", "comercial_bt", "residencial"),
      adotantes = c(11857, 315815, 798073),
      mercado_potencial = c(16718, 412539, 3370592),
      total_ucs = c(181394, 12098612, 72212363),
      penetracao_total = c(0.06536600, 0.02610341, 0.01105175),
      mercado_nicho = c(181394, 1930557, 11522806),
      penetracao_nicho = c(0.0653660, 0.1635875, 0.0692603),
      penetracao_potencial = c(0.7092356, 0.7655397, 0.2367753)
    )
)

epe4md_proj_mensal(
  lista_potencia = lista_potencia,
  ano_base = 2021,
  ano_max_resultado = 2050,
  ajuste_ano_corrente = FALSE,
  ultimo_mes_ajuste = NA,
  metodo_ajuste = NA,
  dir_dados_premissas = NA_character_
)
}
