% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_calibra_curva_s.R
\encoding{UTF-8}
\name{epe4md_calibra_curva_s}
\alias{epe4md_calibra_curva_s}
\title{Calibra o modelo de Bass com dados históricos e gera curvas S de adoção.}
\usage{
epe4md_calibra_curva_s(
  resultado_payback,
  consumidores,
  ano_base,
  ano_max_resultado = 2060,
  p_max = 0.01,
  q_max = 1,
  curva_bateria = "propria",
  p_bat = 0.0015,
  q_bat = 0.3,
  inicio_curva_bateria = 2023,
  dir_dados_premissas = NA_character_
)
}
\arguments{
\item{resultado_payback}{data.frame. Resultado da função
\link{epe4md_payback}.}

\item{consumidores}{list. Resultado da função
\link{epe4md_mercado_potencial}.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{ano_max_resultado}{numeric. Ano final para apresentação dos resultados.
Máximo igual a 2060. Default igual a 2060.}

\item{p_max}{numeric. Fator de inovação (p) máximo. Default igual a 0.01.}

\item{q_max}{numeric. Fator de imitação (q) máximo. Default igual a 1.}

\item{curva_bateria}{string. Pode ser "propria" (Default) ou "lag". A primeira
opção utiliza os fatores p_bat e q_bat. A opção "lag" utiliza a curva de
difusão de MMGD, atrasada no tempo, conforme o parâmetro inicio_curva_bateria.}

\item{p_bat}{numeric. Fator de inovação (p) para baterias. Default igual a 0.0015.}

\item{q_bat}{numeric. Fator de imitação (q) para_baterias. Default igual a 0.3.}

\item{inicio_curva_bateria}{numeric. Indica o ano que começa a difusão
das baterias. Default igual a 2023.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas. Se esse
parâmetro não for passado, a função usa os dados default que são instalados
com o pacote. É importante que os nomes dos arquivos sejam os mesmos da
pasta default.}
}
\value{
data.frame com curvas de difusão e mercado potencial
}
\description{
Calibra o modelo de Bass com dados históricos e gera curvas S de adoção.
}
