% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_copia_premissas.R
\encoding{UTF-8}
\name{epe4md_copia_premissas}
\alias{epe4md_copia_premissas}
\title{Copia as planilhas de premissas originais do modelo para um novo diretório
chamado dados_premissas, na raiz do projeto atual.}
\usage{
epe4md_copia_premissas(ano_base)
}
\arguments{
\item{ano_base}{numeric. Ano base das premissas que serão copiadas.}
}
\value{
data.frame. Base tratada e resumida.
}
\description{
Copia as planilhas de premissas originais do modelo para um novo diretório
chamado dados_premissas, na raiz do projeto atual.
}
