% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_geracao_ano}
\alias{epe4md_graf_geracao_ano}
\title{Gráfico da geração anual em MWmed}
\usage{
epe4md_graf_geracao_ano(
  dados,
  ano_inicio = 2013,
  cor = "#953735",
  tamanho = 14
)
}
\arguments{
\item{dados}{data.frame Resultados mensais de potencia e energia}
}
\description{
Gráfico da geração anual em MWmed
}
