% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_geracao_mes}
\alias{epe4md_graf_geracao_mes}
\title{Gráfico da geração mensal em MWmed}
\usage{
epe4md_graf_geracao_mes(
  dados,
  ano_inicio = 2013,
  cor = "#953735",
  tamanho = 14
)
}
\arguments{
\item{dados}{data.frame. Resultados mensais de potência e energia.}

\item{ano_inicio}{numeric. Define o ano em que se inicia o eixo x. Default igual a 2013.}

\item{cor}{string. Escolha da cor do gráfico. Default igual a Vermelho.}

\item{tamanho}{numeric. Altera o tamanho da fonte da legenda. Default igual a 14.}
}
\description{
Gráfico de área que mostra a soma da geração de energia em MWmed agrupada por mês.
}
\value{
Plota um gráfico de área que mostra a soma da geração de energia em MWmed agrupada por mês.
}
\examples{
epe4md_graf_geracao_mes(
  dados,
  ano_inicio = 2013,
  cor = "#953735",
  tamanho = 14
)
}
