% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_part_fonte_geracao}
\alias{epe4md_graf_part_fonte_geracao}
\title{Gráfico da participação de cada tecnologia na geração de energia}
\usage{
epe4md_graf_part_fonte_geracao(dados, cor = "#953735", tamanho = 14)
}
\arguments{
\item{dados}{data.frame Resultados mensais de potencia e energia}
}
\description{
Gráfico da participação de cada tecnologia na geração de energia
}
