% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_part_fonte_geracao}
\alias{epe4md_graf_part_fonte_geracao}
\title{Gráfico da participação de cada tecnologia na geração de energia}
\usage{
epe4md_graf_part_fonte_geracao(dados, cor = "#953735", tamanho = 14)
}
\arguments{
\item{dados}{data.frame. Resultados mensais de potência e energia.}

\item{cor}{string. Escolha da cor do gráfico. Default igual a Vermelho.}

\item{tamanho}{numeric. Altera o tamanho da fonte da legenda. Default igual a 14.}

}
\description{
Gráfico de barras que mostra a participação de cada fonte ('Fotovoltaica', 'Termelétrica', 'Hidro', 'Eólica')
na geração de energia somente no parâmetro \code{ano_max_resultado} passado na função \code{epe4md::epe4md_calcula}.
}
\value{
plota um gráfico de barras que mostra a participação de cada fonte ('Fotovoltaica', 'Termelétrica', 'Hidro', 'Eólica')
na geração de energia somente no parâmetro \code{ano_max_resultado} passado na função \code{epe4md::epe4md_calcula}.
}
\examples{
epe4md_graf_part_fonte_geracao(
  dados,
  cor = "#953735",
  tamanho = 14
)
}
