% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_part_fonte_potencia}
\alias{epe4md_graf_part_fonte_potencia}
\title{Gráfico da participação de capacidade por fonte}
\usage{
epe4md_graf_part_fonte_potencia(
  dados,
  ano_inicio = 2013,
  cor = "#112446",
  tamanho = 14
)
}
\arguments{
\item{dados}{data.frame Resultados mensais de potencia e energia}
}
\value{
Gráfico da participação de capacidade por fonte
}
\description{
Gráfico da participação de capacidade por fonte
}
