% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_part_segmento}
\alias{epe4md_graf_part_segmento}
\title{Gráfico da evolução da participação por segmento}
\usage{
epe4md_graf_part_segmento(dados, ano_inicio = 2013, tamanho = 14)
}
\arguments{
\item{dados}{data.frame Resultados mensais de potencia e energia}
}
\description{
Gráfico da evolução da participação por segmento
}
