% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_part_segmento}
\alias{epe4md_graf_part_segmento}
\title{Gráfico da evolução da participação por segmento}
\usage{
epe4md_graf_part_segmento(dados, ano_inicio = 2013, tamanho = 14)
}
\arguments{
\item{dados}{data.frame. Resultados mensais de potência e energia.}

\item{ano_inicio}{numeric. Define o ano em que se inicia o eixo x. Default igual a 2013.}

\item{tamanho}{numeric. Altera o tamanho da fonte da legenda. Default igual a 14.}
}
\description{
Gráfico de colunas que mostra a participação de cada segmento 'Comercial (AT)', 'Comercial(BT)', 'Comercial Remoto (AT/BT)', Residencial' e 'Residencial Remoto' na geração de energia em cada ano.
}
\value{
Plota um gráfico de colunas que mostra a participação de cada segmento 'Comercial (AT)', 'Comercial(BT)', 'Comercial Remoto (AT/BT)', Residencial' e 'Residencial Remoto' na geração de energia em cada ano.
}
\examples{

dados <- tibble::tibble(
  data = 2021-01-01,
  ano = 2021,
  mes = 1,
  nome_4md = c("ENEL RJ", "ENF", "LIGHT"),
  subsistema = "SE",
  uf = "RJ",
  segmento = "comercial_at",
  fonte_resumo = "Fotovoltaica",
  energia_mwh = c(2019.62250, 69.06433, 2303.18754),
  energia_autoc_mwh = c(1615.69800, 55.25146, 1842.55003),
  energia_inj_mwh = c(403.92450, 13.81287, 460.63751),
  energia_mwmed = c(2.71454638, 0.09282839, 3.09568218),
  pot_mes_mw =  c(1.5275, 0.0000, 1.0500),
  adotantes_mes = c(21, 0, 4),
  p = c(0.0002782611, 0.0006078518, 0.0013757006),
  q = c(1.0000000, 1.0000000, 0.5282319)
)

epe4md_graf_part_segmento(
  dados = dados,
  ano_inicio = 2021,
  tamanho = 14
)
}
