% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_pot_anual}
\alias{epe4md_graf_pot_anual}
\title{Gráfico da capacidade instalada anual}
\usage{
epe4md_graf_pot_anual(dados, ano_inicio = 2013, cor = "#13475d", tamanho = 14)
}
\arguments{
\item{dados}{data.frame Resultados mensais de potencia e energia}
}
\value{
Gráfico da capacidade instalada anual
}
\description{
Gráfico da capacidade instalada anual
}
