% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_pot_cenario}
\alias{epe4md_graf_pot_cenario}
\title{Gráfico da capacidade instalada acumulada por cenário}
\usage{
epe4md_graf_pot_cenario(dados, ano_inicio = 2013, tamanho = 14)
}
\arguments{
\item{dados}{data.frame Resultado resumido com coluna cenario}
}
\value{
Gráfico da capacidade instalada acumulada por cenário
}
\description{
Gráfico da capacidade instalada acumulada por cenário
}
