% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_graficos.R
\encoding{UTF-8}
\name{epe4md_graf_pot_regiao}
\alias{epe4md_graf_pot_regiao}
\title{Gráfico da capacidade instalada acumulada por região}
\usage{
epe4md_graf_pot_regiao(dados, ano_inicio = 2013, tamanho = 14)
}
\arguments{
\item{dados}{data.frame Resultados mensais de potencia e energia}
}
\description{
Gráfico da capacidade instalada acumulada por região
}
