% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_investimentos.R
\encoding{UTF-8}
\name{epe4md_investimentos}
\alias{epe4md_investimentos}
\title{Calcula o montante investido}
\usage{
epe4md_investimentos(
  resultados_mensais,
  ano_base,
  ano_max_resultado = 2050,
  dir_dados_premissas = "inst/dados_premissas"
)
}
\arguments{
\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{ano_max_resultado}{numeric. Ano final para apresentação dos resultados.
Máximo igual a 2050. Default igual a 2050.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas.
Se esse parâmetro não for passado, a função usa os dados default que são
instalados com o pacote. É importante que os nomes dos arquivos sejam os
mesmos da pasta default.}

\item{resultados_mensais.}{Resultado da função \link{epe4md_calcula}.}
}
\value{
data.frame com o montante estimado de investimentos relativos à
expansão da capacidade instalada de micro e minigeração distribuída.
}
\description{
Calcula o montante investido
}
