% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_payback.R
\encoding{UTF-8}
\name{epe4md_payback}
\alias{epe4md_payback}
\title{Roda um fluxo de caixa para cada caso e retorna métricas financeiras.}
\usage{
epe4md_payback(
  casos_payback,
  premissas_reg,
  ano_base,
  sequencial,
  altera_sistemas_existentes = TRUE,
  ano_decisao_alteracao = 2023,
  inflacao = 0.0375,
  taxa_desconto_nominal = 0.13,
  custo_reforco_rede = 200,
  ano_troca_inversor = 11,
  pagamento_disponibilidade = 0.3,
  disponibilidade_kwh_mes = 100,
  desconto_capex_local = 0,
  anos_desconto = 0,
  dir_dados_premissas = NA_character_
)
}
\arguments{
\item{casos_payback}{data.frame. Base gerada pela função
\link{epe4md_casos_payback}}

\item{premissas_reg}{data.frame. Input de premissas regulatórias para serem
consideradas nos cálculos.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{altera_sistemas_existentes}{logic. TRUE se alterações regulatórias
afetam investimentos realizados em anos anteriores à revisão da regulação.
Default igual a FALSE.}

\item{ano_decisao_alteracao}{numeric. Ano em que são definidas novas regras e
se tornam de conhecimento público. Esse parâmetro só tem efeito caso o
anterior seja igual a TRUE. Default igual a 2021.}

\item{inflacao}{mumeric. Taxa anual de inflacao considerada no reajuste das
tarifas e para calcular o retorno real de projetos. Default igual a 0.0375.}

\item{taxa_desconto_nominal}{numeric. Taxa de desconto nominal considerada
nos cálculos de payback descontado. Default igual a 0.13.}

\item{custo_reforco_rede}{numeric. Custo em R$/kW aplicado a projetos de
geracao remota em Alta Tensão. Representa um custo pago pelo empreendedor
para reforços na rede. Default igual a 200.}

\item{ano_troca_inversor}{numeric. Ano, a partir do ano de instalação, em que
é realizada a troca do inversor fotovoltaico. Default igual a 11.}

\item{pagamento_disponibilidade}{numeric. Percentual de meses em que o
consumidor residencial paga custo de disponbilidade em função da
variabilidade da geração FV. Default igual a 0.3. Tem efeito somente até o
ano de 2022.}

\item{disponibilidade_kwh_mes}{numeric. Consumo de disponbilidade do
consumidor em kWh/mês. Default igual a 100, equivalente a um consumidor
trifásico. Tem efeito somente até o ano de 2022.}

\item{desconto_capex_local}{numeric. Percentual de desconto a ser aplicado no
CAPEX de sistemas de geração local(ex: 0.1) para simulação de incentivos.
Default igual a 0.}

\item{anos_desconto}{vector. Anos em que há a incidência do desconto no
CAPEX. Default igual a 0.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas. Se esse
parâmetro não for passado, a função usa os dados default que são instalados
com o pacote. É importante que os nomes dos arquivos sejam os mesmos da
pasta default.}
}
\value{
data.frame. Métricas financeiras para cada caso.
}
\description{
Roda um fluxo de caixa para cada caso e retorna métricas financeiras.
}
