% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_proj_adotantes.R
\encoding{UTF-8}
\name{epe4md_proj_adotantes}
\alias{epe4md_proj_adotantes}
\title{Realiza a projeção do número de adotantes de micro e minigeração distribuída}
\usage{
epe4md_proj_adotantes(
  casos_otimizados,
  consumidores,
  ano_base,
  dir_dados_premissas = NA_character_
)
}
\arguments{
\item{casos_otimizados}{data.frame. Resultado da funcao
\link{epe4md_calibra_curva_s}.}

\item{consumidores}{list. Resultado da função
\link{epe4md_mercado_potencial}.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas. Se esse
parâmetro não for passado, a função usa os dados default que são instalados
com o pacote. É importante que os nomes dos arquivos sejam os mesmos
da pasta default.}
}
\value{
list com dois data.frames. "proj_adotantes" possui os resultados da
projeção de adotantes de micro e minigeração distribuída. "part_adotantes"
possui o resultado em termos de participação do número de adotantes frente
ao total de unidades consumidoras.
}
\description{
Realiza a projeção do número de adotantes de micro e minigeração distribuída
}
\examples{

casos_otimizados <- tibble::tibble(
    nome_4md = c("OUTRA", "RGE", "RORAIMA", "SULGIPE", "UHENPAL"),
    segmento = "comercial_at",
    p = c(0.0004484362, 0.0005572423, 0.0001223451, 0.0002427430, 0.0001034085),
    q = 1,
    spb = 0.3,
    ano = 2021,
    Ft = c(0.7847763, 0.8193315, 0.4980477, 0.6633618, 0.4560837),
    consumidores = c(2801, 10995, 791, 116, 48),
    payback = c(9.300698, 7.295061, 11.896997, 12.448897, 11.590019),
    mercado_potencial = c(172, 1232, 22, 3, 1)
)

consumidores <- tibble(
    nome_4md = c("CELESC", "CEMIG", "COELBA", "COPEL", "CPFL PAULISTA"),
    ano = 2021,
    segmento = "residencial",
    consumidores = c(294853, 705996, 328907, 488333, 531478)
)

epe4md_proj_adotantes(
  casos_otimizados = casos_otimizados,
  consumidores = consumidores,
  ano_base = 2021,
  dir_dados_premissas = NA_character_
)
}
