% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_proj_adotantes.R
\encoding{UTF-8}
\name{epe4md_proj_adotantes}
\alias{epe4md_proj_adotantes}
\title{Realiza a projeção do número de adotantes de micro e minigeração distribuída}
\usage{
epe4md_proj_adotantes(
  casos_otimizados,
  consumidores,
  ano_base,
  cresc_fv = 0,
  cresc_eol = 0,
  cresc_cgh = 0,
  cresc_ute = 0,
  dir_dados_premissas = NA_character_
)
}
\arguments{
\item{casos_otimizados}{data.frame. Resultado da funcao
\link{epe4md_calibra_curva_s}.
#' @param consumidores list. Resultado da função
\link{epe4md_mercado_potencial}.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{cresc_fv}{numeric. Taxa de crescimento da participação da fonte
fotovoltaica. Default igual a 0 (mantém participação histórica).}

\item{cresc_eol}{numeric. Taxa de crescimento da participação da fonte
eólica. Default igual a 0 (mantém participação histórica).}

\item{cresc_cgh}{numeric. Taxa de crescimento da participação da fonte
hidráulica. Default igual a 0 (mantém participação histórica).}

\item{cresc_ute}{numeric. Taxa de crescimento da participação da fonte
termelétrica. Default igual a 0 (mantém participação histórica).}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas. Se esse
parâmetro não for passado, a função usa os dados default que são instalados
com o pacote. É importante que os nomes dos arquivos sejam os mesmos
da pasta default.}
}
\value{
list com dois data.frames. "proj_adotantes" possui os resultados da
projeção de adotantes de micro e minigeração distribuída. "part_adotantes"
possui o resultado em termos de participação do número de adotantes frente
ao total de unidades consumidoras.
}
\description{
Realiza a projeção do número de adotantes de micro e minigeração distribuída
}
\examples{
\dontrun{
# Exemplo básico

casos_otimizados <- epe4md_calibra_curva_s(
  resultado_payback = epe4md_payback(ano_base = 2023),
  consumidores = epe4md_mercado_potencial(ano_base = 2023),
  ano_base = 2023
)

consumidores <- epe4md_mercado_potencial(ano_base = 2023)

resultado_proj <- epe4md_proj_adotantes(
  casos_otimizados = casos_otimizados,
  consumidores = consumidores,
  ano_base = 2023,
  cresc_fv = 0.05,   # Exemplo: crescimento anual 5\% na fonte fotovoltaica
  cresc_eol = 0.02   # Exemplo: crescimento anual 2\% na fonte eólica
)

head(resultado_proj$proj_adotantes)
head(resultado_proj$part_adotantes)
}
}
