% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_proj_adotantes.R
\encoding{UTF-8}
\name{epe4md_proj_adotantes}
\alias{epe4md_proj_adotantes}
\title{Realiza a projeção do número de adotantes de micro e minigeração distribuída}
\usage{
epe4md_proj_adotantes(
  casos_otimizados,
  consumidores,
  ano_base,
  dir_dados_premissas = NA_character_
)
}
\arguments{
\item{casos_otimizados}{data.frame. Resultado da funcao
\link{epe4md_calibra_curva_s}.}

\item{consumidores}{list. Resultado da função
\link{epe4md_mercado_potencial}.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas. Se esse
parâmetro não for passado, a função usa os dados default que são instalados
com o pacote. É importante que os nomes dos arquivos sejam os mesmos
da pasta default.}
}
\value{
list com dois data.frames. "proj_adotantes" possui os resultados da
projeção de adotantes de micro e minigeração distribuída. "part_adotantes"
possui o resultado em termos de participação do número de adotantes frente
ao total de unidades consumidoras.
}
\description{
Realiza a projeção do número de adotantes de micro e minigeração distribuída
}
