% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_proj_geracao.R
\encoding{UTF-8}
\name{epe4md_proj_geracao}
\alias{epe4md_proj_geracao}
\title{Estima a geração de eletricidade a partir da projeção de potência}
\usage{
epe4md_proj_geracao(proj_mensal, ano_base, dir_dados_premissas = NA_character_)
}
\arguments{
\item{proj_mensal}{dataframe. Resultado da função
\link{epe4md_proj_mensal}.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas.
Se esse parâmetro não for passado, a função usa os dados default que são
instalados com o pacote. É importante que os nomes dos arquivos sejam os
mesmos da pasta default.}
}
\value{
data.frame com os resultados da projeção de capacidade instalada
de micro e minigeração distribuída, número de adotantes e geração
mensal de energia.
}
\description{
Estima a geração de eletricidade a partir da projeção de potência
}
