% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_proj_geracao.R
\encoding{UTF-8}
\name{epe4md_proj_geracao}
\alias{epe4md_proj_geracao}
\title{Estima a geração de eletricidade a partir da projeção de potência}
\usage{
epe4md_proj_geracao(
  proj_mensal,
  ano_base,
  filtro_de_uf,
  dir_dados_premissas = "inst/dados_premissas"
}
\arguments{
\item{proj_mensal}{dataframe. Resultado da função
\link{epe4md_proj_mensal}.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{filtro_de_uf}{string. Parâmetro que define uma unidade federativa (UF) a
ser filtrada. Caso uma UF não seja indicada ou seja informado um valor inválido,
o resultado será apresentado sem filtros.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas.
Se esse parâmetro não for passado, a função usa os dados default que são
instalados com o pacote. É importante que os nomes dos arquivos sejam os
mesmos da pasta default.}
}
\value{
data.frame com os resultados da projeção de capacidade instalada
de micro e minigeração distribuída, número de adotantes e geração
mensal de energia.
}
\description{
Estima a geração de eletricidade a partir da projeção de potência
}
\examples{

  projecao_mensal <- structure(
    list(ano = c(2015, 2025, 2022),
         nome_4md = c("COPEL", "ESE", "UHENPAL"),
         fonte_resumo = c("Fotovoltaica", "Fotovoltaica", "Hidro"),
         segmento = c("comercial_at", "residencial", "residencial"),
         adotantes_mes = c(1, 172, 0),
         pot_mes_mw = c(0.00767, 1.14501518259998, 0),
         mes_ano = structure(c(16467, 20240, 19236),
                             class = c("yearmonth", "vctrs_vctr")),
         mes = c(2, 6, 9),
         p = c(0.000505391053085753, 0.000786787452721811, 0.00364169896575091),
         q = c(0.872731853324977, 0.543083860089691, 0.560319141756969),
         Ft = c(0.00731468279517835, 0.629706876035949, 0.64417085772448)),
    row.names = c(NA, -3L),
    class = c("tbl_df", "tbl", "data.frame"))

  epe4md_proj_geracao(
    proj_mensal = projecao_mensal,
    ano_base = 2021
  )
}

