% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_proj_mensal.R
\encoding{UTF-8}
\name{epe4md_proj_mensal}
\alias{epe4md_proj_mensal}
\title{Faz a abertura mensal da projeção de potência}
\usage{
epe4md_proj_mensal(
  lista_potencia,
  ano_base,
  ano_max_resultado = 2050,
  ajuste_ano_corrente = FALSE,
  ultimo_mes_ajuste = NA,
  metodo_ajuste = NA,
  dir_dados_premissas = NA_character_
)
}
\arguments{
\item{lista_potencia}{list. Resultado da função
\link{epe4md_proj_potencia}.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{ano_max_resultado}{numeric. Ano final para apresentação dos resultados.
Máximo igual a 2050. Default igual a 2050.}

\item{ajuste_ano_corrente}{logic. Se TRUE indica que a projeção deverá
incorporar o histórico mensal recente, verificado em parte do primeiro ano
após o ano base. Default igual a FALSE. O arquivo base_mmgd.xlsx deve
incorporar esse histórico.}

\item{ultimo_mes_ajuste}{numeric. Último mês com dados completos na
base_ano_corrente. Default igual a NA. Só tem efeito caso ajuste_ano_corrente
seja igual a TRUE.}

\item{metodo_ajuste}{string. Se igual a "extrapola" o modelo irá extrapolar a
potência e o número de adotantes até o final do ano base + 1 com base no
verificado até o ultimo_mes_ajuste. Se igual a "substitui", o modelo
substitui a projeção até o ultimo_mes_ajuste e mantém o restante do ano com a
projeção normal. Só tem efeito caso ajuste_ano_corrente seja igual a TRUE.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas.
Se esse parâmetro não for passado, a função usa os dados default que são
instalados com o pacote. É importante que os nomes dos arquivos sejam os
mesmos da pasta default.}
}
\value{
data.frame com os resultados da projeção de capacidade instalada
de micro e minigeração distribuída e número de adotantes mensal
}
\description{
Faz a abertura mensal da projeção de potência
}
\examples{
  lista_potencia <- list(
    structure(
      list(nome_4md = c("SULGIPE", "ENEL RJ", "COSERN"),
           segmento = c("residencial_remoto", "residencial", "comercial_at"),
           p = c(0.000323066811243587, 0.000123056130181494, 0.000787809069463831),
           q = c(0.721680688512525, 0.875693111591128, 1),
           spb = c(0.3, 0.3, 0.3),
           ano = c(2025, 2025, 2019),
           Ft = c(0.842113510009708,  0.925209402868019, 0.464451352457598),
           consumidores = c(1114, 286430, 2714),
           payback = c(3.64182621961061, 3.6477830110581, 7.44160219022217),
           mercado_potencial = c(93.5, 23971.5, 72.75),
           adotantes_ano = c(44, 10982, 0),
           adotantes_acum = c(309, 88861, 1),
           fonte_resumo = c("Fotovoltaica", "Fotovoltaica", "Eólica"),
           part_fonte = c(1, 1, 0.0028169014084507),
           adotantes_hist = c(NA, NA, 0),
           pot_media = c(2.16376623376623, 5.52175275167785, 100),
           pot_ano = c(95.2057142857143, 60639.8887189262, 0),
           pot_hist = c(NA, NA, 0),
           pot_ano_mw = c(0.0952057142857143, 60.6398887189262, 0),
           pot_acum_mw = c(0.668603766233766, 485.066631017987, 0.1)),
      class = c("tbl_df", "tbl", "data.frame"),
      row.names = c(NA, -3L)),
    structure(
      list(ano = c(2023, 2013, 2027),
           segmento = c("comercial_bt", "comercial_at", "comercial_at"),
           adotantes = c(457589, 6, 28934),
           mercado_potencial = c(484014, 117, 28385),
           total_ucs = c(12524434, 180897, 199518),
           penetracao_total = c(0.0365357029307672, 3.31680459045755e-05, 0.14501949698774),
           mercado_nicho = c(2002218, 180897, 199518),
           penetracao_nicho = c(0.228541047977793, 3.31680459045755e-05, 0.14501949698774),
           penetracao_potencial = c(0.945404471771478, 0.0512820512820513, 1.01934120133874)),
      row.names = c(NA, -3L),
      class = c("tbl_df", "tbl", "data.frame")))

  epe4md_proj_mensal(
    lista_potencia = lista_potencia,
    ano_base = 2021
  )
}
