% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_proj_potencia.R
\encoding{UTF-8}
\name{epe4md_proj_potencia}
\alias{epe4md_proj_potencia}
\title{Realiza a projecao da capacidade instalada de micro e minigeracao distribuida}
\usage{
epe4md_proj_potencia(
  lista_adotantes,
  ano_base,
  dir_dados_premissas = "inst/dados_premissas"
)
}
\arguments{
\item{lista_adotantes}{data.frame. Data.frame que contém dados relativos aos
adotantes.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas.
Se esse parâmetro não for passado, a função usa os dados default que são
instalados com o pacote. É importante que os nomes dos arquivos sejam os
mesmos da pasta default.}

\item{proj_adotantes}{data.frame. Resultado da função}
}
\value{
list com dois data.frames. "proj_potencia" possui os resultados da
projeção de capacidade instalada de micro e minigeração distribuída.
"part_adotantes" possui o resultado em termos de participação do
número de adotantes frente ao total de unidades consumidoras.
}
\description{
\link{epe4md_proj_adotantes}.
}
