% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_proj_potencia.R
\encoding{UTF-8}
\name{epe4md_proj_potencia}
\alias{epe4md_proj_potencia}
\title{Realiza a projeção da capacidade instalada de micro e minigeração distribuída}
\usage{
epe4md_proj_potencia(
  lista_adotantes,
  ano_base,
  dir_dados_premissas = NA_character_
)
}
\arguments{
\item{lista_adotantes}{list. Resultado da função
\link{epe4md_proj_adotantes}.}

\item{ano_base}{numeric. Ano base da projeção. Define o ano em que a função
irá buscar a base de dados. Último ano completo realizado.}

\item{dir_dados_premissas}{Diretório onde se encontram as premissas.
Se esse parâmetro não for passado, a função usa os dados default que são
instalados com o pacote. É importante que os nomes dos arquivos sejam os
mesmos da pasta default.}

\item{perc_pot_bateria}{numeric. Percentual de potência atribuída a baterias.
Default igual a 0.35 (35\%).}
}
\value{
list com dois data.frames. "proj_potencia" possui os resultados da
projeção de capacidade instalada de micro e minigeração distribuída.
"part_adotantes" possui o resultado em termos de participação do
número de adotantes frente ao total de unidades consumidoras.
}
\description{
Realiza a projeção da capacidade instalada de micro e minigeração distribuída
}
\examples{
\dontrun{
# Supondo que você já tenha rodado as funções anteriores para obter a lista de adotantes:
lista_adotantes <- epe4md_proj_adotantes(
  casos_otimizados = casos_otimizados,
  consumidores = consumidores,
  ano_base = 2023
)

resultado_potencia <- epe4md_proj_potencia(
  lista_adotantes = lista_adotantes,
  ano_base = 2023
)

head(resultado_potencia$proj_potencia)
head(resultado_potencia$part_adotantes)
}
}
