% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epe4md_calcula.R
\name{epe4md_sumariza_resultados}
\alias{epe4md_sumariza_resultados}
\title{Resume os resultados de capacidade instalada}
\usage{
epe4md_sumariza_resultados(resultados_mensais)
}
\arguments{
\item{resultados_mensais}{data.frame. Saída da função
\code{epe4md::epe4md_calcula}}
}
\value{
data.frame com projeção de capacidade instalada nacional, em GW e
geração de energia, em GWh e MWméd.
}
\description{
Resume os resultados de capacidade instalada
}
\examples{
\dontrun{
resultados <- epe4md_calcula(premissas_reg, ano_base = 2023)
resumo <- epe4md_sumariza_resultados(resultados)
}

}
